% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.officer_get_all_layout_properties}
\alias{.officer_get_all_layout_properties}
\title{Get all layouts' properties}
\usage{
.officer_get_all_layout_properties(pptx_obj)
}
\arguments{
\item{pptx_obj}{\code{officer::rpptx} object}
}
\value{
An object of class \code{data.frame} with fields for placeholder
  attributes and one row per placeholder element per layout.
}
\description{
Gets all the layout properties for all the layouts in an officer presentation
}
\keyword{internal}
