% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_all_layout_properties}
\alias{get_all_layout_properties}
\title{Get all layouts' properties}
\usage{
get_all_layout_properties(x)
}
\arguments{
\item{x}{\code{R2PpptxPresentation} object}
}
\value{
An object of class \code{data.frame} with fields for placeholder
  attributes and one row per placeholder element per layout.
}
\description{
Gets all the layout properties for all the layouts in a presentation
}
