% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_layout_properties}
\alias{get_layout_properties}
\title{Gets layout properties}
\usage{
get_layout_properties(x, layout)
}
\arguments{
\item{x}{\code{R2PpptxPresentation} object}

\item{layout}{character. Name of the layout to see properties for.}
}
\value{
An object of class \code{data.frame} with fields for placeholder
  attributes and one row per placeholder element.
}
\description{
A thin wrapper around \code{officer::plot_layout_properties()} to get layouts
properties for \code{R2PpptxPresentation} objects.
}
