% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_layouts}
\alias{get_layouts}
\title{Get layouts}
\usage{
get_layouts(x)
}
\arguments{
\item{x}{\code{R2PpptxPresentation} object}
}
\value{
An object of class \code{data.frame} with fields for the layout name
  and the name of the slide theme (master).
}
\description{
A thin wrapper around \code{officer::layout_summary()} to get layouts for
\code{R2PpptxPresentation} objects.
}
