% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slide.R
\name{get_slides,R2PptxSlideList-method}
\alias{get_slides,R2PptxSlideList-method}
\title{method to get slides}
\usage{
\S4method{get_slides}{R2PptxSlideList}(x)
}
\arguments{
\item{x}{R2PptxSlideList object}
}
\description{
method to get slides
}
