% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slide.R, R/presentation.R
\name{length,R2PptxSlide-method}
\alias{length,R2PptxSlide-method}
\alias{length,R2PptxPresentation-method}
\title{get slide length (elements)}
\usage{
\S4method{length}{R2PptxSlide}(x)

\S4method{length}{R2PptxPresentation}(x)
}
\description{
get slide length (elements)

get presentation length (slides)
}
