% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/element.R
\name{location,R2PptxElement-method}
\alias{location,R2PptxElement-method}
\title{Get location (key) of element}
\usage{
\S4method{location}{R2PptxElement}(x)
}
\description{
Get location (key) of element
}
\keyword{internal}
