% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/element.R
\name{new_element,character-method}
\alias{new_element,character-method}
\title{New Element}
\usage{
\S4method{new_element}{character}(key, value)
}
\arguments{
\item{key}{character. Name of the placeholder label for this element.}

\item{value}{object. Object to put into a PowerPoint slide, eg text or a plot.}
}
\value{
An object of class \code{R2PptxElement} representing something to put
  on a slide.
}
\description{
Make a new `R2PptxElement`. Element represent text, a flextable, an image, a
ggplot2, etc to add to a slide.
}
