% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/element.R
\name{new_element}
\alias{new_element}
\title{New Element}
\usage{
new_element(key, value)
}
\description{
Make a new `R2PptxElement`. Element represent text, a flextable, an image, a
ggplot2, etc to add to a slide.
}
\section{Slots}{

\describe{
\item{\code{key}}{character. Name of the placeholder label for this element.}

\item{\code{value}}{object. Object to put into a powerpoint slide, eg text or a plot.}
}}

