% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/location.R
\name{new_location}
\alias{new_location}
\title{New Location}
\usage{
new_location(ph_location_fn, ...)
}
\arguments{
\item{ph_location_fn}{function. Function in the \code{officer::ph_location*} family}

\item{...}{args to pass to \code{ph_location_fn}}
}
\value{
An object of class \code{R2PptxLocation} representing the location to
  place an \code{R2PptxElement}.
}
\description{
New Location
}
\examples{
# create an element with some text offset 2 from the left and 2 from the top
# of the slide
element_location <- new_location(officer::ph_location, left = 2, top = 2)
element <- new_element(
  key = element_location,
  value = "Some text"
)
presentation <- new_presentation() +
  new_slide("Title Slide", elements = list(element))
path <- tempfile(fileext = ".pptx")
write_pptx(presentation, path)
if (interactive()) browseURL(path)
}
\seealso{
\link[officer]{ph_location}
}
