% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/presentation.R
\name{new_presentation}
\alias{new_presentation}
\title{New Presentation}
\usage{
new_presentation(
  template_path = getOption("default_pptx_template"),
  slides = list()
)
}
\arguments{
\item{template_path}{character. Path of the file that has the PowerPoint
template to use. Defaults to path set in `options("default_pptx_template")`}

\item{slides}{list. Optional. List of slides to initiate the presentation
with.}
}
\value{
An object of class \code{R2PptxPresentation} representing a future
  PowerPoint presentation.
}
\description{
Make a new `R2PptxPresentation`. Presentations represent PowerPoint decks.
}
\examples{
# see https://mattle24.github.io/r2pptx/articles/basic_usage.html
# for more documentation
presentation <- new_presentation()
print(presentation)
print(template_path(presentation))

title_slide <- new_slide(
  layout = "Title Slide",
  elements = list(
    new_element(key = "Title 1", value = "The title"),
    new_element(key = "Subtitle 2", value = "The subtitle")
  )
)
print(title_slide)
presentation <- presentation + title_slide
print(presentation)
ppt_path <- tempfile(fileext = ".pptx")
write_pptx(presentation, ppt_path)
if (interactive()) system(paste("open", ppt_path))
}
