% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot_layout}
\alias{plot_layout}
\title{Plot layout}
\usage{
plot_layout(x, layout)
}
\arguments{
\item{x}{\code{R2PpptxPresentation} object}

\item{layout}{character. Name of the layout to see properties for.}
}
\value{
No return value, called for side effects
}
\description{
A thin wrapper around \code{officer::plot_layout_properties()} to plot
layouts for \code{R2PpptxPresentation} objects.
}
