% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{+,R2Pptx,R2PptxSlide-method}
\alias{+,R2Pptx,R2PptxSlide-method}
\title{Add R2Ppptx slide}
\usage{
\S4method{+}{R2Pptx,R2PptxSlide}(e1, e2)
}
\description{
Add an `R2PptxSlide` object to something else compatible.
}
