% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/presentation.R
\name{+,R2PptxPresentation,R2PptxSlideList-method}
\alias{+,R2PptxPresentation,R2PptxSlideList-method}
\title{Add R2Ppptx slidelist}
\usage{
\S4method{+}{R2PptxPresentation,R2PptxSlideList}(e1, e2)
}
\arguments{
\item{e1}{`R2PptxPresentation` object}

\item{e2}{`R2PptxSlideList` object}
}
\description{
Add an `R2PptxSlideList` object to a presentation.
}
\keyword{internal}
