% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slide.R
\name{+,R2PptxSlide,R2PptxElement-method}
\alias{+,R2PptxSlide,R2PptxElement-method}
\title{Add element to slide}
\usage{
\S4method{+}{R2PptxSlide,R2PptxElement}(e1, e2)
}
\arguments{
\item{e1}{\code{R2PptxSlide} object}

\item{e2}{\code{R2PptxElement} object}
}
\value{
An object of class \code{R2PptxSlide}, which is \code{e1} with an
additional element \code{e2}
}
\description{
Add an \code{R2PptxElement} object to an \code{R2PptxSlide} object.
}
