% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slide.R
\name{+,R2PptxSlide,R2PptxSlideList-method}
\alias{+,R2PptxSlide,R2PptxSlideList-method}
\title{Add slide to slidelist}
\usage{
\S4method{+}{R2PptxSlide,R2PptxSlideList}(e1, e2)
}
\arguments{
\item{e1}{\code{PR2PptxSlide} object}

\item{e2}{\code{R2PptxSlideList} object}
}
\value{
An object of class \code{R2PptxSlideList} which is the
  \code{R2PptxSlideList} list \code{e1} with an additional slide which is
  \code{e2}.
}
\description{
Add slide to slidelist
}
