% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slide.R
\name{+,R2PptxSlideList,R2PptxSlideList-method}
\alias{+,R2PptxSlideList,R2PptxSlideList-method}
\title{Add slidelist to slidelist}
\usage{
\S4method{+}{R2PptxSlideList,R2PptxSlideList}(e1, e2)
}
\arguments{
\item{e1}{\code{R2PptxSlideList} object}

\item{e2}{\code{R2PptxSlideList} object}
}
\value{
An object of class \code{R2PptxSlideList} which is \code{e1} with
  additional slides (all the slides in \code{e2}).
}
\description{
Add slidelist to slidelist
}
