% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/location.R
\docType{class}
\name{R2PptxLocation-class}
\alias{R2PptxLocation-class}
\title{Location}
\description{
A Location is used to place an element on a slide. For now this will just be
a placeholder name, but eventually it could expand to be an arbitrary
location
}
