% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/presentation.R
\docType{class}
\name{R2PptxPresentation-class}
\alias{R2PptxPresentation-class}
\title{Presentation}
\description{
An S4 class to represent a powerpoint presentation.
}
\section{Slots}{

\describe{
\item{\code{slides}}{list. List of `R2PptxSlide` objects.}

\item{\code{template_path}}{character. File path to the Powerpoint template used
for this presentation.}
}}

