% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slide.R
\docType{class}
\name{R2PptxSlide-class}
\alias{R2PptxSlide-class}
\title{Slide}
\description{
An S4 class to represent a PowerPoint slide
}
\section{Slots}{

\describe{
\item{\code{layout}}{character. Name of the PowerPoint layout to use for this
slide.}

\item{\code{elements}}{list. List of `R2PptxElement` objects.}
}}

