% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/presentation.R
\name{template_path}
\alias{template_path}
\alias{template_path,R2PptxPresentation-method}
\title{Get template path}
\usage{
template_path(x)

\S4method{template_path}{R2PptxPresentation}(x)
}
\arguments{
\item{x}{object to get the template path for.}
}
\description{
Get template path
}
\section{Methods (by class)}{
\itemize{
\item \code{R2PptxPresentation}: Get the template path of an `R2PptxPresentation`
object.
}}

