% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/element.R
\name{value,R2PptxElement-method}
\alias{value,R2PptxElement-method}
\title{Get value of element}
\usage{
\S4method{value}{R2PptxElement}(x)
}
\value{
Object to put into a PowerPoint slide, eg text or a plot.
}
\description{
Get value of element
}
\keyword{internal}
