% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/presentation.R
\name{write_pptx}
\alias{write_pptx}
\alias{write_pptx,R2PptxPresentation-method}
\title{Write pptx}
\usage{
write_pptx(x, path)

\S4method{write_pptx}{R2PptxPresentation}(x, path)
}
\arguments{
\item{x}{object}

\item{path}{character. File path to write to.}
}
\value{
An object of class \code{R2PptxPresentation}
}
\description{
Write an object to a `.pptx` file.
}
\section{Methods (by class)}{
\itemize{
\item \code{R2PptxPresentation}: Write a presentation to a `.pptx` file
}}

