% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_layout_properties}
\alias{get_layout_properties}
\title{Gets layout properties}
\usage{
get_layout_properties(x, layout)
}
\arguments{
\item{x}{R2PpptxPresentation object}

\item{layout}{character. Name of the layout to see properties for.}
}
\description{
A thin wrapper around `officer::plot_layout_properties()` to plot
layouts for `R2PpptxPresentation` objects.
}
