% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_layouts}
\alias{get_layouts}
\title{Get layouts}
\usage{
get_layouts(x)
}
\arguments{
\item{x}{R2PpptxPresentation object}
}
\description{
A thin wrapper around `officer::layout_summary()` to get layouts for
`R2PpptxPresentation` objects.
}
