% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slide.R
\name{get_slides,R2PptxSlideList-method}
\alias{get_slides,R2PptxSlideList-method}
\title{method to get slides}
\usage{
\S4method{get_slides}{R2PptxSlideList}(x)
}
\arguments{
\item{x}{\code{R2PptxSlideList} object}
}
\value{
List, a list of \code{R2PptxSlide} objects.
}
\description{
method to get slides
}
