% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slide.R, R/presentation.R
\name{length,R2PptxSlide-method}
\alias{length,R2PptxSlide-method}
\alias{length,R2PptxPresentation-method}
\title{get slide length (elements)}
\usage{
\S4method{length}{R2PptxSlide}(x)

\S4method{length}{R2PptxSlide}(x)

\S4method{length}{R2PptxPresentation}(x)
}
\arguments{
\item{x}{`R2PptxSlide` object}
}
\value{
Integer, the number of slides in the presentation.
}
\description{
Returns the number of elements in a slide
}
