% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/element.R
\name{new_element}
\alias{new_element}
\title{New Element}
\usage{
new_element(key, value)
}
\value{
An object of class \code{R2PptxElement} representing something to put
  on a slide.
}
\description{
Make a new `R2PptxElement`. Element represent text, a flextable, an image, a
ggplot2, etc to add to a slide.
}
