% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/presentation.R
\name{new_presentation}
\alias{new_presentation}
\title{New Presentation}
\usage{
new_presentation(
  template_path = getOption("default_pptx_template"),
  slides = list()
)
}
\arguments{
\item{template_path}{character. Path of the file that has the Powerpoint
template to use. Defaults to path set in `options("default_pptx_template")`}

\item{slides}{list. Optional. List of slides to initiate the presentation
with.}
}
\description{
Make a new `R2PptxPresentation`. Presentations represent powerpoint decks.
}
