% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slide.R
\name{new_slide}
\alias{new_slide}
\title{New slide}
\usage{
new_slide(layout, elements = list())
}
\arguments{
\item{layout}{character. Name of the powerpoint layout to use for this
slide.}

\item{elements}{list. List of `R2PptxElements` to initialize the slide with.
Defaults to empty list.}
}
\description{
Make a `R2PptxSlide` object representing a powerpoint slide.
}
