% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slide.R
\name{new_slidelist}
\alias{new_slidelist}
\title{New slide list}
\usage{
new_slidelist(slides = list())
}
\arguments{
\item{slides}{list. List of \code{R2PptxSlide} objects to initialize the list with.
Defaults to empty list.}
}
\value{
An object of class \code{R2PptxSlideList} representing a list of
  \code{R2pptxSlide} object.
}
\description{
Make a \code{R2PptxSlideList} object representing a list of PowerPoint slides
}
