% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slide.R
\name{+,R2PptxSlide,R2PptxElement-method}
\alias{+,R2PptxSlide,R2PptxElement-method}
\title{Add element to slide}
\usage{
\S4method{+}{R2PptxSlide,R2PptxElement}(e1, e2)
}
\description{
Add an `R2PptxElement` object to an `R2PptxSlide` object.
}
