% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slide.R
\name{+,R2PptxSlide,R2PptxSlideList-method}
\alias{+,R2PptxSlide,R2PptxSlideList-method}
\title{Add slide to slidelist}
\usage{
\S4method{+}{R2PptxSlide,R2PptxSlideList}(e1, e2)
}
\description{
Add slide to slidelist
}
