% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/location.R
\docType{class}
\name{R2PptxLocation-class}
\alias{R2PptxLocation-class}
\title{Location}
\description{
A Location is used to place an element on a slide.
}
\section{Slots}{

\describe{
\item{\code{ph_location_fn}}{function. Function in the \code{officer::ph_location*} family}

\item{\code{args}}{list. Arguments to pass to \code{ph_location_fn}}
}}

