% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/location.R
\docType{class}
\name{R2PptxLocation-class}
\alias{R2PptxLocation-class}
\title{Placeholder Location}
\description{
A Location is used to place an element on a slide. This class handles
locations that are the names of PowerPoint placeholders.
}
