% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slide.R
\docType{class}
\name{R2PptxSlideList-class}
\alias{R2PptxSlideList-class}
\title{Slide list}
\description{
\code{R2PptxSlideList} is an S4 class to contain groups of \code{R2PptxSlide}
objects that are not part of a presentation. It is meant to be used to create
lists of slides and then be able to add the list easily to a presentation.
}
\section{Slots}{

\describe{
\item{\code{slides}}{list. A list of \code{R2PptxSlide} objects}
}}

