% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/presentation.R
\name{template_path<-}
\alias{template_path<-}
\alias{template_path<-,R2PptxPresentation-method}
\title{Set template path}
\usage{
template_path(x) <- value

\S4method{template_path}{R2PptxPresentation}(x) <- value
}
\arguments{
\item{x}{object to set the template path of.}

\item{value}{character. File path of the new template}
}
\description{
Set template path
}
\section{Functions}{
\itemize{
\item \code{template_path<-,R2PptxPresentation-method}: Set the template path of an `R2PptxPresentation`
object.
}}

